kaggle_requirements.txt
# Please keep this in alphabetical order
Boruta
Cartopy
ImageHash
Janome
PyArabic
PyUpSet
Pympler
Rtree
shapely
SimpleITK
TPOT
Wand
bayesian-optimization
boto3
catboost
category-encoders
cesium
comm
cytoolz
# Older versions of datasets fail with "Loading a dataset cached in a LocalFileSystem is not supported"
# https://stackoverflow.com/questions/77433096/notimplementederror-loading-a-dataset-cached-in-a-localfilesystem-is-not-suppor
datasets>=2.14.6
deap
dipy
docker
easyocr
emoji
fasttext
featuretools
fiona
fury
fuzzywuzzy
geojson
gensim
# b/443054743,b/455550872
google-adk[a2a,eval]>=1.21.0
google-cloud-aiplatform
google-cloud-videointelligence
google-cloud-vision
google-genai
gpxpy
h2o
haversine
hep-ml
igraph
ipympl
ipywidgets==8.1.5
isoweek
jedi
# jitler 0.11.1 breaks simulation image
jiter==0.10.0
# b/276358430: fix Jupyter lsp freezing up the jupyter server
jupyter-lsp==1.5.1
# b/333854354: pin jupyter-server to version 2.12.5; later versions break LSP (b/333854354)
jupyter_server==2.12.5
jupyter_server_proxy
jupyterlab
jupyterlab-lsp
# b/409363708: Ensure we have the update version, we can consider removing it once
# Colab base image is updated more frequently.
kaggle>=1.7.4.2
kaggle-environments
keras-cv
keras-nlp
keras-tuner
kornia
langid
# b/328788268: libpysal 4.10 seems to fail with "module 'shapely' has no attribute 'Geometry'. Did you mean: 'geometry'"
libpysal<=4.9.2
lime
line_profiler
mamba
matplotlib
mlcrate
mne
mpld3
# b/274619697: learntools requires a specific nbconvert right now
nbconvert==6.4.5
nbdev
nilearn
olefile
onnx
openslide-bin
openslide-python
optuna
pandas-profiling
pandasql
papermill
path
path.py
pdf2image
plotly-express
preprocessing
pudb
pyLDAvis
pycryptodome
pydicom
pyemd
pyexcel-ods
pymongo
pypdf
pytesseract
python-lsp-server
pytorch-ignite
pytorch-lightning
qgrid
qtconsole
ray
rgf-python
s3fs
scikit-learn
# Scikit-learn accelerated library for x86
scikit-learn-intelex>=2023.0.1
scikit-multilearn
scikit-optimize
scikit-plot
scikit-surprise
git+https://github.com/facebookresearch/segment-anything.git
squarify
tensorflow-cloud
tensorflow-io
torchinfo
torchmetrics
torchtune
transformers>=4.51.0
vtk
wavio
xvfbwrapper
ydata-profiling
